backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
cadquery==2.4.0
anthropic==0.18.0
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.26.3
